pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = "Document-based RAG chatbot"
authors = ["bonnie <jhryu@bistelligence.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.1"
python-multipart = "^0.0.20"
pymupdf = "^1.25.5"
openai = "^1.74.0"
langchain = "^0.3.23"
faiss-cpu = "^1.10.0"
tiktoken = "^0.9.0"
langchain-community = "^0.3.21"
langchain-core = "^0.3.51"
sentence-transformers = "^4.1.0"
langchain-huggingface = "^0.1.2"
pypdf = "^5.4.0"
langchain-ollama = "^0.3.1"
sse-starlette = "^2.3.3"
langchain-openai = "^0.3.14"
pandas = "^2.2.3"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
